backend/requirements.txt
# Backend — ClinicMonitoring (pinlar CI va Docker uchun barqarorlik)
Django>=5.0,<6
whitenoise>=6.6,<7
djangorestframework>=3.15,<4
django-cors-headers>=4.3,<5
channels>=4.0,<5
daphne>=4.0,<5
channels-redis>=4.2,<5
dj-database-url>=2.1,<3
psycopg2-binary>=2.9,<3
google-generativeai>=0.8,<1
Pillow>=10,<12
python-dotenv>=1.0,<2
